+++ src/buffer.h
@@ -46,7 +46,7 @@ protected:
    bool save;  // save skipped data
    int save_max;
 
-   long pos;
+   off_t pos;
 
    // low-level for derived classes
    virtual int Put_LL(const char *buf,int size) { return 0; }
@@ -77,11 +77,11 @@ public:
    void PutEOF() { eof=true; PutEOF_LL(); }
 
    // useful for cache.
-   void Save(long m) { save=true; save_max=m; }
+   void Save(int m) { save=true; save_max=m; }
    void GetSaved(const char **buf,int *size) const;
-   void SaveRollback(long p);
+   void SaveRollback(int p);
 
-   void SetPos(long p) { pos=p; }
+   void SetPos(off_t p) { pos=p; }
 
    void Empty();
